requirements.txt
# Core requirements
PyQt6
numpy
pyqtgraph
python-dateutil

# MQTT for using MQTT mode
paho-mqtt

# The below depend on the modules used
pyvisa
pyserial
scipy
nidaqmx
matplotlib